requirements.txt
cryptography
django-jazzmin
django-debug-toolbar
Pillow
pytesseract
langchain
python-dotenv
dj-database-url
langchain_community
langchain_huggingface
num2words